play queue is atom now
