+++ web/src/Components/Search.js
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faSpotify } from "@fortawesome/free-brands-svg-icons";
-import { useQueue } from "../hooks/useQueue";
+import { useQueue } from "../hooks/use-queue";
 import { useSearch } from "../hooks/use-search";
 
 export default function Search() {
